refactor(quiz) rename page components to more descriptive names
- Rename LoginPage → AuthLoginPage (update file paths, imports, stories, and tests)
- Rename CreateUserPage → AuthRegisterPage (update file paths, imports, stories, and tests)
- Rename DiscoverPage → QuizDiscoverPage (update file paths, imports, stories, and tests)
- Rename JoinPage → GameJoinPage (update file paths, imports, stories, and tests)
- Update router configuration to use new component names
