+++ packages/quiz/src/pages/AuthLoginPage/components/AuthLoginPageUI/index.ts
@@ -1,0 +1,2 @@
+export type { AuthLoginPageUIProps, LoginFormFields } from './AuthLoginPageUI'
+export { default } from './AuthLoginPageUI'
