+++ packages/quiz/src/pages/AuthLoginPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AuthLoginPage'
