+++ packages/quiz/src/pages/AuthRegisterPage/AuthRegisterPage.tsx
@@ -1,0 +1,26 @@
+import React, { FC, useState } from 'react'
+import { useNavigate } from 'react-router-dom'
+
+import { useQuizServiceClient } from '../../api/use-quiz-service-client.tsx'
+
+import { AuthRegisterPageUI, CreateUserFormFields } from './components'
+
+const AuthRegisterPage: FC = () => {
+  const navigate = useNavigate()
+
+  const { login, register } = useQuizServiceClient()
+
+  const [isLoading, setIsLoading] = useState<boolean>(false)
+
+  const handleSubmit = (values: CreateUserFormFields) => {
+    setIsLoading(true)
+    register(values)
+      .then(() => login({ email: values.email, password: values.password }))
+      .then(() => navigate('/'))
+      .finally(() => setIsLoading(false))
+  }
+
+  return <AuthRegisterPageUI loading={isLoading} onSubmit={handleSubmit} />
+}
+
+export default AuthRegisterPage
