+++ packages/quiz/src/pages/AuthRegisterPage/components/AuthRegisterPageUI/helpers.ts
