+++ packages/quiz/src/pages/AuthRegisterPage/components/AuthRegisterPageUI/index.ts
@@ -1,0 +1,5 @@
+export type {
+  AuthRegisterPageUIProps,
+  CreateUserFormFields,
+} from './AuthRegisterPageUI'
+export { default } from './AuthRegisterPageUI'
