+++ packages/quiz/src/pages/AuthRegisterPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AuthRegisterPage'
