--- packages/quiz/src/pages/CreateUserPage/components/CreateUserPageUI/CreateUserPageUI.stories.tsx
@@ -1,23 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-import { withRouter } from 'storybook-addon-remix-react-router'
-
-import CreateUserPageUI from './CreateUserPageUI'
-
-const meta = {
-  title: 'Pages/CreateUserPage',
-  component: CreateUserPageUI,
-  decorators: [withRouter],
-  parameters: {
-    layout: 'fullscreen',
-  },
-} satisfies Meta<typeof CreateUserPageUI>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const Default = {
-  args: {
-    loading: false,
-    onSubmit: () => undefined,
-  },
-} satisfies Story
