--- packages/quiz/src/pages/CreateUserPage/components/CreateUserPageUI/index.ts
@@ -1,5 +1,0 @@
-export type {
-  CreateUserPageUIProps,
-  CreateUserFormFields,
-} from './CreateUserPageUI'
-export { default } from './CreateUserPageUI'
