--- packages/quiz/src/pages/CreateUserPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CreateUserPage'
