--- packages/quiz/src/pages/DiscoverPage/components/DiscoverPageUI/index.ts
@@ -1,2 +1,0 @@
-export type { DiscoverPageUIProps } from './DiscoverPageUI'
-export { default } from './DiscoverPageUI'
