--- packages/quiz/src/pages/DiscoverPage/components/index.ts
@@ -1,1 +1,0 @@
-export { default as DiscoverPageUI } from './DiscoverPageUI'
