--- packages/quiz/src/pages/DiscoverPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DiscoverPage'
