+++ packages/quiz/src/pages/GameJoinPage/GameJoinPage.stories.tsx
@@ -1,0 +1,21 @@
+import type { Meta, StoryObj } from '@storybook/react'
+import { withRouter } from 'storybook-addon-remix-react-router'
+
+import { withMockAuth } from '../../../.storybook/mockAuthContext.tsx'
+
+import GameJoinPage from './GameJoinPage'
+
+const meta = {
+  component: GameJoinPage,
+  decorators: [withRouter, withMockAuth],
+  parameters: {
+    layout: 'fullscreen',
+  },
+} satisfies Meta<typeof GameJoinPage>
+
+export default meta
+type Story = StoryObj<typeof meta>
+
+export const Default = {
+  args: {},
+} satisfies Story
