+++ packages/quiz/src/pages/GameJoinPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GameJoinPage'
