--- packages/quiz/src/pages/JoinPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './JoinPage'
