--- packages/quiz/src/pages/LoginPage/components/index.ts
@@ -1,2 +1,0 @@
-export type { LoginPageUIProps, LoginFormFields } from './LoginPageUI'
-export { default as LoginPageUI } from './LoginPageUI'
