--- packages/quiz/src/pages/LoginPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from './LoginPage'
