+++ packages/quiz/src/pages/QuizDiscoverPage/components/QuizDiscoverPageUI/index.ts
@@ -1,0 +1,2 @@
+export type { QuizDiscoverPageUIProps } from './QuizDiscoverPageUI'
+export { default } from './QuizDiscoverPageUI'
