+++ packages/quiz/src/pages/QuizDiscoverPage/components/index.ts
@@ -1,0 +1,1 @@
+export { default as QuizDiscoverPageUI } from './QuizDiscoverPageUI'
