+++ packages/quiz/src/pages/QuizDiscoverPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './QuizDiscoverPage'
