+++ packages/quiz/src/pages/index.ts
@@ -1,14 +1,14 @@
 export { default as AuthGamePage } from './AuthGamePage'
-export { default as CreateUserPage } from './CreateUserPage'
-export { default as DiscoverPage } from './DiscoverPage'
+export { default as AuthLoginPage } from './AuthLoginPage'
+export { default as AuthRegisterPage } from './AuthRegisterPage'
 export { default as ErrorPage } from './ErrorPage'
-export { default as ProfileGamesPage } from './ProfileGamesPage'
+export { default as GameJoinPage } from './GameJoinPage'
 export { default as GamePage } from './GamePage'
 export { default as GameResultsPage } from './GameResultsPage'
 export { default as HomePage } from './HomePage'
-export { default as JoinPage } from './JoinPage'
-export { default as LoginPage } from './LoginPage'
+export { default as ProfileGamesPage } from './ProfileGamesPage'
+export { default as ProfileQuizzesPage } from './ProfileQuizzesPage'
 export { default as ProfileUserPage } from './ProfileUserPage'
 export { default as QuizCreatorPage } from './QuizCreatorPage'
 export { default as QuizDetailsPage } from './QuizDetailsPage'
-export { default as ProfileQuizzesPage } from './ProfileQuizzesPage'
+export { default as QuizDiscoverPage } from './QuizDiscoverPage'
